resource_server/requirements.txt
django
django-extensions
django-cors-headers
django-sslserver
PyJWT
requests
gunicorn
cryptography
psycopg2-binary
Pillow
django-oauth-toolkit
